Move MiqTask#human_status and constants to module
Meant for future updates to MiqReportResult, and will be used so that
the logic for HumanStatus can be shared with othermodels, as long as
they implment the `state` and `status` methods.

Idealy, we can avoid needing to instanciate MiqTask when fetching the
human_status for a given MiqReportResult.
